Fix crate compilation for different features
Fix crate compilation for different features:
- `bls` with no features.
- `lib` with `metrics-server` feature and without features.
- `account` primitives with no features.

Also move the `PedersenGenerator` implementation of `beserial`'s
(de)serialization to the serialization file within the `bls` crate
and introduce a `cache` feature to handle properly the `lazy` feature
dependency in the cache code.
