+++ primitives/block/Cargo.toml
@@ -27,7 +27,7 @@ log = { package = "tracing", version = "0.1", features = ["log"] }
 serde = { version = "1.0", features = ["derive"], optional = true }
 
 beserial = { path = "../../beserial", features = ["derive"] }
-nimiq-bls = { path = "../../bls", features = ["beserial"]}
+nimiq-bls = { path = "../../bls", features = ["beserial", "cache"]}
 nimiq-collections = { path = "../../collections" }
 nimiq-database = { path = "../../database" }
 nimiq-handel = { path = "../../handel" }
